Code cleanup: remove unused files and reorganize modules
Phase 1 - Quick Wins:
- Delete unused rendering/sprites.py (~180 LOC)
- Remove debug console.log statements from useWebSocket.ts
- Delete unused frontend/src/assets/react.svg
- Consolidate documentation (remove SEPARATION_README.md)

Phase 2 - File Organization:
- Move movement_strategy.py to core/ directory
- Update all imports to use core.movement_strategy
- Fix type annotation in entity_factory.py (Entity -> Agent)

Cleanup impact:
- Removed 708 lines, added 10 lines
- Better code organization with movement_strategy in core/
- Cleaner documentation structure
- Removed debug logging from production code

All core imports verified successfully.
