+++ core/entities.py
@@ -14,7 +14,7 @@ from core.math_utils import Vector2
 
 if TYPE_CHECKING:
     from core.environment import Environment
-    from movement_strategy import MovementStrategy
+    from core.movement_strategy import MovementStrategy
     from core.ecosystem import EcosystemManager
     from core.genetics import Genome
