+++ core/movement_strategy.py
