+++ tests/test_movement_strategy.py
@@ -5,7 +5,7 @@ from pygame.math import Vector2
 
 from agents import Fish, Crab, Food
 from core.environment import Environment
-from movement_strategy import (MovementStrategy, SoloFishMovement, SchoolingFishMovement,
+from core.movement_strategy import (MovementStrategy, SoloFishMovement, SchoolingFishMovement,
                                 CRAB_AVOIDANCE_DISTANCE, SOLO_FISH_AVOIDANCE_DISTANCE,
                                 SCHOOLING_FISH_ALIGNMENT_DISTANCE)
